Added Savannah and City BattleMats. These are just combos of maps from the respective map pack, but added for completeness sake.
For "Trench" there is a grassland theme and a volcano theme (I feel like the volcano theme matches the tone of the map art better).

"Palace" is heavily interpreted, as there's no real indication on the map what is supposed to be a building and what's just higher terrain. Similar to the City Map Pack, there is a lot of terrain that looks like pavement, but it probably intended to be played as clear. I have included a version where most of the map is paved hexes, and a version where most are clear terrain.